+++ script/python3/sunlight_control.py
@@ -2,8 +2,9 @@
 # -*- coding: utf-8 -*-
 # this made for python3
 
-import os, sys, argparse, yaml, subprocess as sp
-from multiprocessing import Process, Event
+import os, sys, argparse, yaml, math, time, subprocess as sp
+# from multiprocessing import Process, Event
+from threading import Thread, Event
 from datetime import datetime, timedelta
 
 from util.env import expand_env
@@ -11,7 +12,12 @@ from util.timer import LEDLightDayTimer
 from util.remote import Remote
 from util.weather_info import WeatherInfo
 from util import module_logger
-from util.display_ssd1331.as_clock import Screen
+if os.uname().sysname == 'Darwin':
+    print('## the ENV Cannot use luma library ##')
+    from display.dummy_opts import get_device
+else:
+    from luma.core.render import canvas
+    from display.demo_opts import get_device
 
 from time import sleep
 
@@ -22,9 +28,9 @@ SETTING = os.path.normpath(os.path.join(_BASE, '../../settings/ledlight.yml'))
 
 DEBUG = False
 
-class SunlightControl(object):
+class SunlightControl(Thread):
 
-    def __init__(self, timer, per_sec, setting=None):
+    def __init__(self, timer, per_sec, update_event, setting=None):
         if __name__ == '__main__':
             self.ARGS = SunlightControl.ArgParser()
             self.config_path = self.ARGS.configure
@@ -37,6 +43,7 @@ class SunlightControl(object):
           self.PARAMS = expand_env(params, DEBUG)
 
         self._per_sec = per_sec # to check every _perse
+        self._update_event = update_event
         timer.timezone = self.PARAMS['TIMEZONE']
         self.remotes = {}
         # restrict update lircd for runnning
@@ -50,9 +57,12 @@ class SunlightControl(object):
         timer.remote = self.remotes['ledlight']
         self._timer = timer
         self.logger = module_logger(__name__)
+        super().__init__()
 
-        self.stop_event = Event()
-        self._thread = Process(target=self.run, args=())
+    @property
+    def updated(self):
+        assert hasattr(self, '_update_event')
+        return self._update_event
 
     @property
     def timer(self):
@@ -71,20 +81,6 @@ class SunlightControl(object):
     @active_schedules.setter
     def active_schedules(self, val):
         self._active_schedules = val
-
-    @property
-    def thread(self):
-        assert hasattr(self, '_thread')
-        return self._thread
-
-    def start(self):
-        assert hasattr(self, '_thread')
-        self._thread.start()
-        return self
-
-    def kill(self):
-        assert hasattr(self, 'stop_event')
-        self.stop_event.set()
 
     # operate transferred instance
     def _setup_wether_info(self, day):
@@ -112,10 +108,11 @@ class SunlightControl(object):
     def run(self):
         day = datetime.now(self.timer.timezone)
         self.active_schedules = None
-        while not self.stop_event.is_set():
+        while True:
             if self.timer.is_usedup():
                 self.logger.info('Schedules set for day: %s'%day.strftime('%Y-%m-%d'))
                 self.active_schedules = self.update_settings(day)
+                self.updated.set() # new schedules were setuped
                 self.logger.info('Schedules = {}'.format(self.active_schedules))
 
             if len(self.active_schedules) > 0:
@@ -123,7 +120,6 @@ class SunlightControl(object):
             else:
                 # search active day to fetch astronomy data
                 day += timedelta(days=1)
-                continue
 
     @staticmethod
     def ArgParser():
@@ -139,20 +135,101 @@ class SunlightControl(object):
             help='config file that wrote by yaml describe params, see default=%s'%SETTING)
         return argParser.parse_args()
 
-def kill_thread(threading_instance):
-    if threading_instance.thread.isAlive():
-        threading_instance.kill()
+
+class Screen(Thread):
+    def __init__(self, update_event, sunctrl):
+        self._device = get_device(['-d', 'ssd1331', '-i', 'spi', '--width', '96', '--height', '64'])
+        self._update_event = update_event
+        self._sunctrl = sunctrl
+        super().__init__()
+
+    @property
+    def updated(self):
+        assert hasattr(self, '_update_event')
+        return self._update_event
+
+    @property
+    def device(self):
+        assert hasattr(self, '_device')
+        return self._device
+
+    def posn(self, angle, arm_length):
+        dx = int(math.cos(math.radians(angle)) * arm_length)
+        dy = int(math.sin(math.radians(angle)) * arm_length)
+        return (dx, dy)
+
+
+    def run(self):
+        today_last_time = "Unknown"
+        device = self.device
+        an_lineheight = 8
+
+        while True:
+            now = datetime.now()
+            today_time = now.strftime("%H:%M:%S")
+            if today_time != today_last_time:
+                today_last_time = today_time
+                with canvas(device) as draw:
+                    now = datetime.now()
+                    schs = None
+                    if self.updated.is_set():
+                        # SunlightControl.active_schedules is ready
+                        # get most recent schedule for future
+                        schs = ins.active_schedules
+                        self.updateed.clear()
+                    if (schs is not None) and len(schs) > 0:
+                        print('Schedules = {}'.format(schs))
+
+                    today_date = now.strftime("%y%m%d")
+
+                    margin = 4
+
+                    cx = 30
+                    cy = min(device.height, 64) / 2
+
+                    left = cx - cy
+                    right = cx + cy
+
+                    hrs_angle = 270 + (30 * (now.hour + (now.minute / 60.0)))
+                    hrs = self.posn(hrs_angle, cy - margin - 7)
+
+                    min_angle = 270 + (6 * now.minute)
+                    mins = self.posn(min_angle, cy - margin - 2)
+
+                    sec_angle = 270 + (6 * now.second)
+                    secs = self.posn(sec_angle, cy - margin - 2)
+                    # dimension ssd1331 96 x 64
+                    draw.ellipse((left + margin, margin, right - margin, min(device.height, 64) - margin), outline="yellow")
+                    draw.line((cx, cy, cx + hrs[0], cy + hrs[1]), fill="white")
+                    draw.line((cx, cy, cx + mins[0], cy + mins[1]), fill="white")
+                    draw.line((cx, cy, cx + secs[0], cy + secs[1]), fill="cyan")
+                    draw.ellipse((cx - 2, cy - 2, cx + 2, cy + 2), fill="white", outline="white")
+
+                    # literal infos
+                    draw.text((1.8 * (cx + margin), cy - an_lineheight * 4), 'next:', fill="yellow")
+                    draw.text((1.8 * (cx + margin), cy - an_lineheight * 2), today_date, fill="yellow")
+                    draw.text((2 * (cx + margin), cy), today_time, fill="yellow")
+            time.sleep(0.1)
 
 if __name__ == '__main__':
-    ins = None
+    # use like shared flag
+    update = Event()
     try:
-        ins = SunlightControl(LEDLightDayTimer(), 30 * 60).start()
-        sleep(1)
+        # preprocess
+        print('#####')
+
+        ins = SunlightControl(LEDLightDayTimer(), 30 * 60, update)
+        ins.setDaemon(True)
+        ins2 = Screen(update, ins)
+        ins2.setDaemon(True)
+        print('%%%%%')
+        # ins.start()
+        # ins.join(5)
         print('XXX')
-        ins2 = Screen(ins.active_schedules).start()
-        ins.thread.join()
-        ins2.thread.join()
+        ins2.start()
+        ins.start()
+        # ins2.join()
     except KeyboardInterrupt:
-        if ins is not None: kill_thread(ins)
-        if ins2 is not None: kill_thread(ins2)
+        ins.join(1)
+        ins2.join(1)
         print('Caught KeyboardInterrupt. schedules were cancelled.')
